Oops.

+++ source/joka/stdc.d
@@ -20,13 +20,13 @@ version (WebAssembly) {
 // math.h
 
 float sqrtf(float x);
-float sqrt(double x);
+double sqrt(double x);
 
 float sinf(float x);
-float sin(double x);
+double sin(double x);
 
 float cosf(float x);
-float cos(double x);
+double cos(double x);
 
 // stdlib.h
